aider: Added pytest to dev-packages and created a script to run pytest in Pipfile.
# Aider chat conversation:

USER: I ran this command:

 pipenv run pytest

And got this output:

Error: the command pytest could not be found within PATH or Pipfile's [scripts].

ASSISTANT: It looks like `pytest` is not being recognized. This could be due to a missing installation or an issue with the virtual environment. Let's ensure `pytest` is correctly installed and available in the `Pipfile`.

First, let's make sure `pytest` is listed under `[dev-packages]` in the `Pipfile`.

Pipfile
<source>toml
<<<<<<< SEARCH
pytest = "*"
=======
pytest = "*"
>>>>>>> REPLACE
</source>

Next, let's add a script to the `Pipfile` to run `pytest`:

Pipfile
<source>toml
<<<<<<< SEARCH
=======
[scripts]
test = "pytest"
>>>>>>> REPLACE
</source>

Now, you can run the tests using:

```bash
pipenv run test
```

This should ensure that `pytest` is correctly recognized and executed within the `pipenv` environment.