feat: Update views to include margin-right in inline-block style
This commit modifies the `app.py` file to update the inline-block style of the graphs in the `update_views` function. It adds a `margin-right` property to the style, with a value of `20px`. This change introduces a margin between the graphs, improving the visual separation and readability of the content.

Signed-off-by: Marcel Hild <hild@b4mad.net>